fix 1

+++ modules/task_2/kolesnikov_i_Cannon_dense_matrix/main.cpp
@@ -3,17 +3,17 @@
 #include "./matrix.h"
 
 TEST(CannonTest, test_1) {
-    size_t size = 16;
+    int size = 16;
     Matrix matrix1(size);
     Matrix matrix2(size);
     std::vector< std::vector<double>> res_matrix;
-    size_t block_size = size/4;
-    size_t block_count = size / block_size;
+    int block_size = size/4;
+    int block_count = size / block_size;
 
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         std::vector<double> vec(size);
         res_matrix.push_back(vec);
-        for (size_t j = 0; j < size; ++j) {
+        for (int j = 0; j < size; ++j) {
             res_matrix[i][j] = 0;
         }
     }
@@ -33,17 +33,17 @@ TEST(CannonTest, test_1) {
 
 
 TEST(CannonTest, test_2) {
-    size_t size = 4;
+    int size = 4;
     Matrix matrix1(size);
     Matrix matrix2(size);
     std::vector< std::vector<double>> res_matrix;
-    size_t block_size = size/2;
-    size_t block_count = size / block_size;
+    int block_size = size/2;
+    int block_count = size / block_size;
 
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         std::vector<double> vec(size);
         res_matrix.push_back(vec);
-        for (size_t j = 0; j < size; ++j) {
+        for (int j = 0; j < size; ++j) {
             res_matrix[i][j] = 0;
         }
     }
@@ -61,17 +61,17 @@ TEST(CannonTest, test_2) {
 }
 
 TEST(CannonTest, test_3) {
-    size_t size = 4;
+    int size = 4;
     Matrix matrix1(size);
     Matrix matrix2(size);
     std::vector< std::vector<double>> res_matrix;
-    size_t block_size = size/2;
-    size_t block_count = size / block_size;
+    int block_size = size/2;
+    int block_count = size / block_size;
 
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         std::vector<double> vec(size);
         res_matrix.push_back(vec);
-        for (size_t j = 0; j < size; ++j) {
+        for (int j = 0; j < size; ++j) {
             res_matrix[i][j] = 0;
         }
     }
@@ -89,17 +89,17 @@ TEST(CannonTest, test_3) {
 }
 
 TEST(CannonTest, test_4) {
-    size_t size = 4;
+    int size = 4;
     Matrix matrix1(size);
     Matrix matrix2(size);
     std::vector< std::vector<double>> res_matrix;
-    size_t block_size = size/2;
-    size_t block_count = size / block_size;
+    int block_size = size/2;
+    int block_count = size / block_size;
 
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         std::vector<double> vec(size);
         res_matrix.push_back(vec);
-        for (size_t j = 0; j < size; ++j) {
+        for (int j = 0; j < size; ++j) {
             res_matrix[i][j] = 0;
         }
     }
@@ -114,23 +114,23 @@ TEST(CannonTest, test_4) {
 
     std::vector<std::vector<double>> matr1 = matrix3.get_matrix(), matr2 = matrix4.get_matrix();
 
-    for (size_t j(0); j < size; ++j) {
+    for (int j(0); j < size; ++j) {
         ASSERT_DOUBLE_EQ(matr1[0][j], matr2[0][j]);
     }
 }
 
 TEST(CannonTest, test_5) {
-    size_t size = 9;
+    int size = 9;
     Matrix matrix1(size);
     Matrix matrix2(size);
     std::vector< std::vector<double>> res_matrix;
-    size_t block_size = size/3;
-    size_t block_count = size / block_size;
+    int block_size = size/3;
+    int block_count = size / block_size;
 
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         std::vector<double> vec(size);
         res_matrix.push_back(vec);
-        for (size_t j = 0; j < size; ++j) {
+        for (int j = 0; j < size; ++j) {
             res_matrix[i][j] = 0;
         }
     }
@@ -145,8 +145,8 @@ TEST(CannonTest, test_5) {
 
     std::vector<std::vector<double>> matr1 = matrix3.get_matrix(), matr2 = matrix4.get_matrix();
 
-    for (size_t i(0); i < size; ++i) {
-        for (size_t j(0); j < size; ++j) {
+    for (int i(0); i < size; ++i) {
+        for (int j(0); j < size; ++j) {
             ASSERT_DOUBLE_EQ(matr1[i][j], matr2[i][j]);
         }
     }
