+++ modules/task_2/kolesnikov_i_Cannon_dense_matrix/matrix.h
@@ -7,20 +7,20 @@
 
 class Matrix {
  public:
-    explicit Matrix(size_t size) :size(size) {
+    explicit Matrix(int size) :size(size) {
         matrix.reserve(size);
-        for (size_t i(0); i < size; ++i) {
+        for (int i(0); i < size; ++i) {
             std::vector<double> vec(size);
             matrix.push_back(vec);
         }
     }
-    Matrix(std::vector<std::vector<double>> matrix, size_t size) :size(size) {
+    Matrix(std::vector<std::vector<double>> matrix, int size) :size(size) {
         this->matrix.reserve(size);
-        for (size_t i(0); i < size; ++i) {
+        for (int i(0); i < size; ++i) {
             std::vector<double> vec;
             this->matrix.push_back(vec);
         }
-        for (size_t i(0); i < size; ++i) {
+        for (int i(0); i < size; ++i) {
             this->matrix[i].assign(matrix[i].begin(), matrix[i].end());
         }
     }
@@ -30,22 +30,22 @@ class Matrix {
     }
     void generateMatrix(double num);
     std::vector< std::vector<double>> cannonAlgorithmSeq(Matrix matrix2,
-    std::vector< std::vector<double>> res_matrix, size_t block_size,
-    size_t block_count);
+    std::vector< std::vector<double>> res_matrix, int block_size,
+    int block_count);
     std::vector< std::vector<double>> cannonAlgorithmOMP(Matrix matrix2,
-    size_t thread_nums, std::vector< std::vector<double>> res_matrix,
-    size_t block_size, size_t block_count);
-    void shiftLeft(std::vector< std::vector<double>> *matr, size_t pos,
-    size_t block_count, size_t skew);
-    void shiftUp(std::vector< std::vector<double>> *matr, size_t pos, size_t block_count,
-    size_t skew);
+    int thread_nums, std::vector< std::vector<double>> res_matrix,
+    int block_size, int block_count);
+    void shiftLeft(std::vector< std::vector<double>> *matr, int pos,
+    int block_count, int skew);
+    void shiftUp(std::vector< std::vector<double>> *matr, int pos, int block_count,
+    int skew);
     void mutiplyByBlock(std::vector< std::vector<double>> block1,
     std::vector< std::vector<double>> block2,
     std::vector< std::vector<double>>* res_block,
-    size_t shift_l, size_t shift_r, size_t skew);
+    int shift_l, int shift_r, int skew);
 
  private:
-    size_t size;
+    int size;
     std::vector<std::vector<double>> matrix;
 };
 #endif  // MODULES_TASK_2_KOLESNIKOV_I_CANNON_DENSE_MATRIX_MATRIX_H_
